feat: Add RoleTestPage
Add a page for testing out how role access works. It uses the AuthorizeView tag in the razor file; the cascaded AuthenticationState; and a RoleManager. The RoleManager lists all available roles, while the AuthenticationState lists the user and can check their name and if they are in a given role.