feat: Add functionality to add new user to /Users
Add the popup dialog to the /Users page. This dialog allows a new user to be created with a specified email and role.

Add buttons to edit and delete users. While the delete button works, the edit button (which is supposed to only allow changing user role) does not properly save at the moment.

Edit the RoleTestPage code-behind to use the same method of user access as /Users.