fix: Make popup able to edit users
Setting the value of the label option in GenericDropdown to (TValue)default meant that changing to the default value was not being picked up by the OnChanged event. The logic for this component is changed by carefully using nullables to track updates and display the menu label.

Change /Users in several ways to reflect this difference. Now instead of the role being a string, it is a nullable LMSRole and must be handled accordingly.